Implement std::hash and equality for graphics pipeline options
GraphicsPipelineOptions is then immediately useful as a key for
pipeline caches.

std::hash also needs size_t return type rather than uint64_t.
MacOS + clang also complaining about this.

Fix compilation of macOS in the CI with the fix to OpenXRInstance
for a -Wswitch compilation error. Must have been enabled in the
CI recently.

Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/154361
Tested-by: Continuous Integration <build@kdab.com>

+++ src/KDGpu/handle.h
@@ -75,7 +75,7 @@ template<typename T>
 struct hash<KDGpu::Handle<T>> {
     size_t operator()(const KDGpu::Handle<T> &handle) const
     {
-        uint64_t hash = 0;
+        size_t hash = 0;
 
         KDGpu::hash_combine(hash, handle.index());
         KDGpu::hash_combine(hash, handle.generation());
