+++ src/KDGpu/utils/hash_utils.h
@@ -16,7 +16,7 @@
 namespace KDGpu {
 
 template<class T>
-inline void hash_combine(uint64_t &seed, const T &v)
+inline void hash_combine(size_t &seed, const T &v)
 {
     std::hash<T> hasher;
     seed ^= hasher(v) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
